Backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
langchain
langchain-google-genai
sqlalchemy
firebase-admin
google-genai